Original change: https://android-review.googlesource.com/c/platform/external/perfetto/+/2864529

+++ ui/src/tracks/process_summary/process_summary_track.ts
@@ -131,9 +131,11 @@ export class ProcessSummaryTrackController extends
   }
 
   async onDestroy(): Promise<void> {
-    await this.query(`drop table if exists ${
-        this.tableName(
-            'window')}; drop table if exists ${this.tableName('span')}`);
+    if (this.engine.isAlive) {
+      await this.engine.query(`drop table if exists ${
+          this.tableName(
+              'window')}; drop table if exists ${this.tableName('span')}`);
+    }
   }
 }
